+++ lib/modules/security/security_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: security
-// Updated at: 2026-04-25 11:38:09
-// Commit: test(security): add unit tests for security business logic
+// Updated at: 2026-04-25 15:57:56
+// Commit: feat(security): add bi-directional event bus handler for security
 
 class securityComponentState {
-  final String commitRef = '662';
-  final String lastMessage = 'test(security): add unit tests for security business logic';
-  final DateTime timestamp = DateTime.parse('2026-04-25T11:38:09+05:00');
+  final String commitRef = '672';
+  final String lastMessage = 'feat(security): add bi-directional event bus handler for security';
+  final DateTime timestamp = DateTime.parse('2026-04-25T15:57:56+05:00');
 }